creating paths file on clone, persistence file structure changed, user meta_dir file

+++ src/main/java/de/arthurpicht/meta/cli/persistence/project/MetaPathsFile.java
@@ -1,4 +1,4 @@
-package de.arthurpicht.meta.tasks.clone;
+package de.arthurpicht.meta.cli.persistence.project;
 
 import de.arthurpicht.meta.cli.target.Target;
 import de.arthurpicht.meta.config.MetaConfig;
@@ -11,18 +11,21 @@ import java.util.List;
 
 public class MetaPathsFile {
 
-    private static final String RepoPathsFileName = "meta-paths";
+    private static final String RepoPathsFileName = "paths";
 
     private final List<String> pathStringList;
     private final Path metaDir;
 
     public MetaPathsFile(MetaConfig metaConfig, Target target) {
-        this.pathStringList = new MetaPathFileInitializer(metaConfig, target).getPathStringList();
+        this.pathStringList = new MetaPathsFileInitializer(metaConfig, target).getPathStringList();
         this.metaDir = metaConfig.getMetaDir();
     }
 
     public void write() throws IOException {
-        Path metaPathFile = this.metaDir.resolve(RepoPathsFileName);
+        if (!PersistenceProjectLocalDir.exists(this.metaDir)) {
+            throw new RuntimeException("Meta project not initialized yet. Local persistence dir not found: [" + PersistenceProjectLocalDir.asPath(this.metaDir) + "].");
+        }
+        Path metaPathFile = PersistenceProjectLocalDir.asPath(this.metaDir).resolve(RepoPathsFileName);
         Files.deleteIfExists(metaPathFile);
         Files.write(metaPathFile, this.pathStringList, StandardOpenOption.CREATE_NEW);
     }
